feat: Introduce new modal API
To better follow the design system
we have introduced a new modal API
to populate the modal-footer slot. We
are deprecating the use of the modal-ok and
modal-cancel slots. Eventually these two slots
will be removed and the only way to populate the
modal footer will be through props. Props such as
ok-title and cancel-title will be deprciated as well.