feat: add emoji to button
Emojis should follow the design pattern laid out,
which is that they are to the left of the text and have
a right margin of 8px.

They are also 16px in a normal button and 12px in
a small button.

They utilize the `selected` attribute to designate
the user has already chosen that emoji.