feat(sprintf): Allow full sentence interpolation
This teaches `GlSprintf` to understand `Start` and `End` placeholders
within a message, as is [common][1] when marking up a translatable
sentence, and pass the content between those placeholders to the scoped
slot of the base name of those placeholders.

This fixes the [outstanding limitation][2] of GlSprintf that string
concatenation and `v-html` otherwise could partially (but inferiorly)
solve, allowing full marked-up sentences (with arbitrary Vue components)
to be translatable.

This should also preserve all existing behaviours of `GlSprintf`,
including edge-cases, and so should be safe to upgrade to.

[1]: https://docs.gitlab.com/ee/development/i18n/externalization.html#avoid-splitting-sentences-when-adding-links
[2]: https://gitlab.com/gitlab-org/gitlab/issues/21344#note_221778896