chore: remove 0 variables
They are needlessly verbose and provide no benefit
0 is exempt from changes to our scale, our spacing, or our
preferred units. 0 transcends language, preprocessor, and
postprocessor.

The only known case (to me) where you want an alternate
value of 0 is for the flex-basis value of flex shorthand
declarations, where you must use 0% for IE10 & 11, however
none of these variables are for that, and are therefore
probably fine to remove.