fix(button): center text in block button
We expect the text to be centered when the
button is full width. This fix is to ensure that
is the case.

However, the dropdown button expects the
opposite to happen. Instead, we want the text
and the caret to be on opposing sides of the
button. We also handle this use case.