ci: Ensure that the security scanning jobs run on master
Apparently the caching job which we have, has a side effect. If we do
not update the cache, the job becomes a manual job. Unfortunately this
blocks the security jobs in the next stage which wait for the manual job
to finish.

We can shortcircuit this by utilizing the empty DAG dependency
functionality.

For reference:
- https://gitlab.com/gitlab-org/gitlab-ui/-/merge_requests/1325
- https://gitlab.com/gitlab-org/gitlab-ui/-/merge_requests/1312