feat(css): Add CSS grid rules
We have no existing CSS grid rules in our codebase. This makes using
CSS grid more complicated as it becomes a per-use styling rather than
something more generalistic.

This adds the structure and some common CSS grid styles that are
already in use. Further styles can be added as and when we need them.