fix: Change generate_utilities.js to handle important in sub rules
This attempts to identify nested include rules and make sure the
important rule is applied correctly. To do this, it walks through all
the inner nodes and when it finds an include rule, it checks against
known declarations and applies the important call if it is found.

If the rule is unknown we throw a console warning to let the developer
know as any includes from other files won't be known.