fix(GlSearchBoxByType): Fix double input event
Prior to this change, `GlSearchBoxByType` maintained its own local copy
of the model value, which had a setter that emitted an `input` event
when it changed. In addition, it forwarded all listeners to the
underlying `GlFormInput`, which emits its _own_ `input` event as well.

These two facts together meant that when a consumer explicitly added
a listener (either via `@input` or `v-model`), it would be called
_twice_: once by the computed setter's `input` event, and again by the
`GlFormInput`'s `input` event.

This change removes the unnecessary local copy of the model, and
explicitly forwards the `input` event from the underlying `GlFormInput`.

Addresses https://gitlab.com/gitlab-org/gitlab-ui/-/issues/937.