fix(GlDropdown): Set default category to primary
In https://gitlab.com/gitlab-org/gitlab-ui/-/merge_requests/1805,
styling of the `tertiary` category for dropdowns was implemented. Prior
to that, the `tertiary` category's styling was equivalent to the
`primary` category's styling.

But, since the default category for the [dropdown button][dd] was
`tertiary`, that meant that all dropdowns which didn't explicitly set
the `category` to `primary` changed in appearance.

Pajamas [currently specifies][pj] that a button's default category is
`tertiary`, but `GlButton` [defaults][btn] to `primary`. So, this change
makes `GlDropdown` agree with `GlButton`, but _not_ Pajamas. This
discrepancy will be dealt with in another MR.

[dd]: https://gitlab.com/gitlab-org/gitlab-ui/-/blob/v24.3.0/src/components/base/dropdown/dropdown.vue#L65
[btn]: https://gitlab.com/gitlab-org/gitlab-ui/-/blob/v24.3.0/src/components/base/button/button.vue#L24
[pj]: https://gitlab.com/gitlab-org/gitlab-services/design.gitlab.com/blob/6e6a7b6a6d5ef8ff8ab574495f43fbefd3e8866e/contents/components/button.md#L41