feat(GlTabs): add action buttons prop
In GitLab a pattern has emerged where 1 ore more buttons are placed
in a specific relation with the Tabs. Currently the group of button
is duplicated to account for mobile and desktop cases.

By utilizing the new props for actions button the need to duplicate
the button(s) is removed. GlTabs component will correctly display
the button(s) in the correct place based on the viewport.