fix(Modal): Autofocus close button, if it is the only focusable element
If there is a modal that has no focusable element in the modal content
AND no modal footer, it should focus the close button.

The close button is the last focusable element we try to focus, and it
doesn't work correctly because it is moved to the end of the list with
the help of Array.splice which doesn't return the Element, but an array
of elements.