ci(GlTabs): disable storyshots on indeterminate story
The `with scroll and growing` tabs story's content relies on a timer
that causes visual tests to be flaky as there might be more or less tabs
in the canvas dependeing on how long the tests are running.

Given that this particular story doesn't provide much value over the
others in terms of visual testing, we can go ahead and disable
storyshots here.