feat: allow unsafe links via the isUnsafeLink prop
This creates a SafeLinkMixin that installs the SafeLinkDirective and
exposes a isUnsafeLink prop. This new mixin is applied to GlLink and
GlButton.

This change lets us render unsafe link with either component when
necessary. Specifically, this aims at skipping the href sanitization
step when rendering download links with Data URLs.