ci: Allow update_screenshots to push to forks
We can push to forks directly, as long as they have "Allow commits from
members who can merge to the target branch". It should be as simple as
changing the remote to the correct project:

https://docs.gitlab.com/ee/user/project/merge_requests/allow_collaboration.html

The workflow is super simple on the maintainer/reviewers side:

Once `Run pipeline` is used to run a pipeline in context of the project,
you can simply run the `update_screenshots` job manually.