chore(deps): update dependency @gitlab/eslint-plugin to v9
This change does not fix the various chart types, as they are more
complicated, and their test coverage isn't sufficient to be confident
about changing them. They can be fixed later.

Also, there's a bug in the interaction between our GlPopover wrapper
component and the underlying BPopover component: naively replacing
`slot="title"` syntax with `#title` breaks title rendering. To work
around this, we need to only set up the title slot in GlPopover if the
consumer of GlPopover is passing that slot. We do this by iterating over
`$scopedSlots`/`$slots`, rather than statically declaring it. The same
fix could be achieved with something like:

    <template v-if="'title' in $slots" #title>

But the iteration approach is more future-proof.