fix(GlFilteredSearch): Stop component mutating `value` prop
GlFilteredSearch mutates the `value` prop. If this prop is
used by the parent component to reactively trigger API
requests, then modifying GlFilteredSearch tokens and text
triggers a lot of unnecessary requests. This change
stops GlFilteredSearch from mutating the prop so that
an API request is triggered only when the user submits
the search query by pressing Enter.